+++ src/base/Component.php
@@ -7,6 +7,8 @@ namespace putyourlightson\sprig\base;
 
 use Craft;
 use craft\base\Component as BaseComponent;
+use craft\helpers\Json;
+use craft\helpers\StringHelper;
 
 abstract class Component extends BaseComponent implements ComponentInterface
 {
@@ -208,9 +210,11 @@ abstract class Component extends BaseComponent implements ComponentInterface
      */
     public static function triggerEvents(array|string $events, string $on = 'load'): void
     {
-        if (is_array($events)) {
-            $events = json_encode(array_combine($events, $events));
+        if (is_string($events)) {
+            $events = StringHelper::split($events);
         }
+
+        $events = array_combine($events, $events);
 
         $headerMap = [
             'load' => 'HX-Trigger',
@@ -221,7 +225,37 @@ abstract class Component extends BaseComponent implements ComponentInterface
         $header = $headerMap[$on] ?? null;
 
         if ($header) {
-            Craft::$app->getResponse()->getHeaders()->set($header, $events);
+            self::addTriggerHeaderEvents($header, $events);
         }
+    }
+
+    /**
+     * Triggers a refresh event on the provided selector.
+     * https://htmx.org/headers/hx-trigger/
+     */
+    public static function triggerRefresh(string $selector): void
+    {
+        self::addTriggerHeaderEvents('HX-Trigger', [
+            'refresh' => [
+                'target' => $selector,
+            ],
+        ]);
+    }
+
+    /**
+     * Adds events to a provided trigger header.
+     */
+    private static function addTriggerHeaderEvents(string $header, array $events): void
+    {
+        $headerValue = Craft::$app->getResponse()->getHeaders()->get($header);
+
+        if ($headerValue !== null) {
+            $currentEvents = Json::decode($headerValue);
+            if (is_array($currentEvents)) {
+                $events = $currentEvents + $events;
+            }
+        }
+
+        Craft::$app->getResponse()->getHeaders()->set($header, Json::encode($events));
     }
 }