+++ src/services/ComponentsService.php
@@ -133,7 +133,7 @@ class ComponentsService extends BaseComponent
      *
      * @since 2.6.0
      */
-    public const HTMX_VERSION = '2.0.1';
+    public const HTMX_VERSION = '2.0.2-pre';
 
     /**
      * @var string|null