+++ tests/pest/Feature/ControllerTest.php
@@ -15,6 +15,7 @@ beforeEach(function() {
     Sprig::$core->controllerMap['test'] = TestController::class;
     Craft::$app->view->setTemplateMode(View::TEMPLATE_MODE_SITE);
     Craft::$app->getView()->setTemplatesPath(CRAFT_TEST_PATH . '/templates');
+    Craft::$app->getView()->clear();
 });
 
 test('Render', function() {